Migrate to new github org. NOTE:
This module depends on v4 of the messaging module, which was defined
in a branch of github.com/smarty/messaging, but the branch is old enough
that it appeared before messaging moved to github.com/smarty and so many
import paths in this package still point to
github.com/smartystreets/messages.

So, there's still some cleanup to be done.